Alterado criação de componentes para a classe de componentes 2D para permitir números flutuantes

+++ SimuladorExperimentos/src/br/simulador/plugin/biblioteca/base/BaseAgente.java
@@ -176,13 +176,13 @@ public abstract class BaseAgente implements IAgente
     }
 
     @Override
-    public int retornar_coordenada_X()
+    public double retornar_coordenada_X()
     {
         return this.coordenadaX;
     }
 
     @Override
-    public int retornar_coordenada_Y()
+    public double retornar_coordenada_Y()
     {
         return this.coordenadaY;
     }
